+++ src/main/java/com/toolrent/demo/service/RentData.java
@@ -1,5 +1,5 @@
 package com.toolrent.demo.service;
 
-public record RentData(String toolType, String brand, String dailyCharge, int chargeDays, double preDiscountCharge) {
+public record RentData(String toolType, String brand, String dailyCharge, double chargeDays, double preDiscountCharge) {
 
 }
